fix(base.js): HTMLImportsLoaded event dispatch throws error on IE
When using a  script loader such as require.js or system.js the HTMLImports polyfill will not load correctly on all versions of IE. It will throw a JavaScript error causing the program to crash. This is caused by the use of `CustomEvent` to dispatch the HTMLImportsLoaded event, which results in `Object doesn't support this function`. By explicitly using the older `createEvent` technique, the bug is fixed.